requirements.txt
nltk
autocorrect
sklearn
textblob
spacy
jupyter
keras
gensim
numpy
clean-text
gunicorn==20.0
traceback2==1.4.0
flask_restx
pyyaml
torch
pytorch_pretrained_bert
flask_cors
scikit-learn